[RESTEASY-2196] Fix MediaType cache poisoning with trailing semi-colon
The `MediaTypeHeaderDelegate` holds a static cache mapping string
representations of Media Types (e.g. application/json) to their
`javax.ws.rs.core.MediaType` equivalent. The reduces repeated parsing of
the same, or similar, Media Types; increasing performance of most
service calls by a small amount.

The class boils down to two methods: `parse(String)` and
`toString(MediaType)`. These two methods are not symmetric since the
parse method may receive additional sections in the Media Type string
(e.g. "application/json; charset=utf-8"). However, it is expected
that when passing a bare Media Type within the additional sections, it
should be symmetric.

Hence: `parse("application/json") == MediaType("application", "json");`

The cache introduces a bug, however, if the `parse` method is called
with a Media Type with a trailing semi-colon. This is technically not
valid for a `Content-Type` but it still can exist. When this happens,
the cache is effectively poisoned as `toString(MediaType("application",
"json"))` will now return "application/json;".

The fix here is to use `internalToString` to normalise the type being
put into the `reverseMap`. The `map` has to stay the same.